initial import of cgp
Collectd Graph Panel is a frontend for Collectd written in PHP.

The goal of CGP is to provide an easy-to-use frontend for Collectd,
starting with page that shows an overview of all the hosts you are
managing with Collectd.

In this initial import there is support for the plugins that are default
enabled in Collectd. The supported plugins are located in the plugin
directory.