add unixsock flush support
Via the UnixSock plugin of Collectd it is possible to send commands to the
Collectd daemon. One of the commands is the FLUSH command, which you can
use to let the daemon write cached data to the rrd files.

CGP uses the FLUSH command before generating a graph and before showing
the load information on the main page. In this case the information shown
via CGP is always up-to-date. Also when you have configured the RRDTool
plugin of Collectd with a CacheTimeout setting.

This commit includes code based on functions collectd_identifier and
collectd_flush from php-collection by Bruno Prémont.