show all type instances in one graph
Planned for collectd version 5.0 is to move the type instance of plugins
df and interface to the plugin instance (see source). From version 5.0 the
default is to group data from a type instance OR data source in one graph.
In case of version 4 there are only 2 plugins to exclude from this: df and
interface.

Source: http://collectd.org/wiki/index.php/Plans_for_5.0