Change how refresh rates for graphs are calculated.
Instead of calculating the refresh rate of graphs as an absolute number
of seconds, it is better if longer duration graphs (1 day, 1 week)
refresh at a slower rate then short duration graphs (2 hours, 8 hours).
The simple way to do this is to use an "aggressiveness" value where
higher values result in more frequent refresh rates on the graphs.

Current calculation is:

(seconds per refresh) = (graph with in seconds) / (aggressiveness value)

For a default of 400 aggressiveness on the detail age, a full day graph
(86400s) will be refreshed every 216s (86400/400).  The maximum refresh
rate would be (86400/1000) = 86.4s.  A conservative value of 50 would
result in a graph that only updates every 1728s (28.8m).

For a graph of very short duration (3600s), a setting of 1000 can result
in the graph being updated every 3.6s.  The more default would be
(3600/400) = 9s, and a conservative setting of 50 would be (3600/50) =
72s.