Calculate refresh rate using detail_graphs_refresh_aggressiveness
Note: Also adds a sanity check that the graph width of a detail graph
cannot be less then 60 seconds.

Grabs the s= parameter off the URL to find out how wide the graph is (in
seconds), or uses the default from the configuration file.  Multiply
this value by 1000 and divide by the aggressiveness value to calculate
the time per refresh (milliseconds) for the javascript code.