Remove PHP warning when an unknown host is requested
When using an unknown host in queries, for example when going to:

    http://localhost/cgp/host.php?h=anUnknownHostWhichDoesNotExist

PHP displays following warning right before a line containing "Unknown host":

    Warning: Invalid argument supplied for foreach() in
    /srv/http/cgp/inc/collectd.inc.php on line 93

Fix this warning by making collectd_plugins() return an empty array in such
situation.  As every caller to this function already handles the special cases
of a false-or-empty return value, this removes the warning and only keeps the
message "Unknown host", as expected.