Fix type confusion
index.php expects that collectd_hosts() always return an array for
array_diff. Since an empty array evaluates to FALSE anyway, do not
change the type and just return the empty array.

In base.php, if no files are returned, the loop is never called and
three variables are not initialized. Do it now.