jsrrdgraph: Fix undefined value issue in RrdDataFile
`RRDRRAInfo.pdp_cnt` does not exist. A reference to the pdp_cnt member
in a struct exists, this is available via the `getPdpPerRow()` method,
so use that. (actually, `getMinStep() * getPdpPerRow()` can be
simplified further, see below).

While at it, replace the combination
`RRDRRAInfo.getPdpPerRow() * RRAInfo.pdp_step` by `RRDRRAInfo.getStep()`
since this is the same thing. As `RRDHeader.pdp_step` is copied into
`RRDRRAInfo`, this can also be substituted.