jsrrdgraph: Improve performance for graphs in the future
The RRD file has finer details for more recent data points. When looking
for a match for a graph past the last update ("the future"), a RRA would
be picked which has a step size of 1. As we know that no data points are
available, optimise this and return earlier with the full range set to
NaN.