hello CGP v1
It's finished :o)

Joking aside. It has been requested multiple times. So let's get it over
with. The last version was more then 3.5 years ago. This will be the last
tagged version of CGP. Every commit in the master branch after this
release can be considered as a new release. ;)

Use git and "git pull" to keep up-to-date or download the latest version
here https://github.com/pommi/CGP/archive/master.zip

Closes: #120 #143