Allow warning/failure threshold HRULEs to be rendered on graphs
This adds an option to the default graph type to add HRULEs on graphs
for warning (yellow) and error (red) thresholds at arbitrary levels.

This is currently being used to draw lines on the load average graph
(CPU count and 2x CPU count).

To work, this fixes a bug in the javascript graph renderer related to
HRULEs without legends, and only counts CPUs whose plugin instances are
numeric.